Reward users with a username via pushPayment
At the moment, when a user completes a quiz question, a new invoice is generate from their wallet and immediately paid by the funder wallet. This is inefficient because it puts unnecessary load on LND. We can instead make pushPayments to users that have a username set, whereas for users without a username we can continue with the older invoice approach.