Refactor dealer bot with composed hedging strategy
Dealer wallet will now compose a hedging strategy that will
abstract the exchange, instrument and actual hedging.
Collateral rebalance will have to stay on the wallet or implemented
in a different pattern as the wallet and the strategy are
interdependent on the collateral rebalancing
via the withdrawal execution and deposit address retrival.