Some changes needed to run integration tests in concourse
We are running the docker-compose containers on an external vm hosting
the docker daemon so we need to be able to access the services from the
outside.
The DOCKER_HOST_IP variable (which defaults to localhost) stands in to
make this possible.
For lnd to work tlsextraip needs to be set to the public ip of the host
so that tls verification will work when connecting.