Re-implement username-to-walletId flow through correct repositories
The first implementations of this involved bypassing the User repository
and calling a 'User' query directly inside the Wallets repository which
is incorrect since the Wallets repo should have no knowledge of the User
repository abstraction.

This commit instead implements the username lookup inside the User
repository, passes the user Id to the Accounts repository when the
user's account is fetched from which we can access the wallets and read
the walletId off the selected wallet (first in the array for now).