Revert 'addInvoiceNoAmount..' calling 'addInvoice..' use case
This was previously done to avoid duplicate code but it ended up failing
when a validation check in 'addInvoice..' caught the 0-sats amount as
invalid which we probably want to keep in here.

I instead implemented a helper function for the walletID steps in those
two methods to help with the duplicated-code concern.