Remove fallback case for when payViaRoute fails
Remove the specific fallback for when a route is calculated using a
particular lnd node, but then the node goes offline just before the user
can finish payment.

The time between a successful probe and failing to be able to send the
payment down that route is small, and is a very low probability event.
This is an edge case where we can reduce some complexity by simply
returning an error that prompts the user to try again.