chore: remove unused 'inverseMax' method
This method in the FeeCalculator isn't used anywhere and is
inconsistently implemented since it does not reliably go back to the
original amount used to calculate the max fee.

Since there isn't a place where it's being used to guide how that method
should be implemented, I opted to simply remove it for now.