test(fix): replace new deviceTokens check
This was added since 'notifications' test is comparing the 'sendBalance'
calls against 'activeAccounts', but it turns out a lot more accounts
without deviceTokens values are later added but are never reached
because the sendNotifications array is shorter.

So in summary, once the first n accounts in the activeAccounts array
have deviceTokens then this test will not break.