fix: staging, following kratos refactoring PR
the issue came from the fact twilio have an _version property that was creating issue for mongoose who also have a _version property on it's object.

this was creating an infinite loop

recreating the PhoneMetadata object with just the necessary property avoid the issue.

a good reason for the proper boundaries between services