feat: combine both admin and public endpoint
there is no reason to launch a dedicated admin server. it makes local dev more memory/cpu heavy.

this is what kratos is doing and I think this is a preferrable approach - the only downside is that if there is a crash on the admin, currently it won't affect public users. but it will with this refactoring

right now this is just a test, we have to do some work on the CI before doing a broader refactoring of the /server files.