chore: remove a critical kratos alert
this is trigger on staging because we try to remove the email, even it doesn't exist.

this action is intended in order to make sure the email is not set, as otherwise the mobile e2e test would fail.

the e2e tests assumes the email is not set and set it on its own