demo: add tests to demonstrate multiple txns
This is to demonstrate 2 scenarios where multiple transactions are
create for the same hash.

To start, run the following:
$ cd core/api
$ make reset-deps
$ bats -t test/bats/ln-send-demo.bats

After this, open the file at 'core/api/.demo.log' and put the provided
query into mongo. Be sure to also add the 'sort' attributes.

The results will be 2 separate changesets:
1. A reimbursement txn should be at the top of the list with amoutns far
   less that what you would expect for the actual successful transaction
   and the wrong transaction direction.

2. A failed txn should appear at the top of the list even though one of
   the attempts was successful.