fix(api): re-add 'require.main' check & admin call from main server
This was removed to support using bun as a runner.
(see https://github.com/oven-sh/bun/issues/4818#issuecomment-1713671800)

There are other issues with bun as well for now, so I'm reverting this
until we need to use bun as a runner and the issue is still not fixed in
bun.