refactor: target tests more specifically in Tiltfile
With auto_init, all other resources are brought up in addition to the
targeted tests. With 'enabled_resources' only the tests and their
dependency resources are brought up.

This is required to run the 'core' test suite for example, since that
suite finishes long before the rest of server resources come up.