fix(core): correct test msat imbalance from fee
In our 'reimburseFee' code we always pass the 'roundedUp' value so as to
not lose millisats in our accounting. Lnd accounts with millisats
though, so if there is a millisat fee that rounds down this will throw
off the balance when accounting.

This commit hardcodes a mitigating 850 msat transaction between lnd
outside nodes 1 and 2. We must use these two nodes because rebalancing
lnd1 will throw off our accounting since we don't have accounting for
rebalances implemented as yet. Fees on lnd_outside_1 are used in other
tests (so setting a zero proportional fee was not explored).