chore: add unique id to write-sdl bin script
The issue here is that on first creation 'public-sdl'/'admin-sdl'
detects that there was a change in ':write-sdl' and correctly
executes. On subsequent builds though, the output from ':write-sdl'
does not change and so 'public-sdl' or 'admin-sdl' does not
re-execute.

This fix adds a unique uuid to the output script of ':write-sdl' to
signal that it was re-executed and that there was some build change in
one of its dependencies.