fix(toolchain): use exec_dep when declaring generator
From https://buck2.build/docs/api/build/attrs/#attrsexec_dep:

Takes a target from the user, as a string, and supplies a dependency to the rule. The dependency will transition to the execution platform. Use exec_dep if you plan to execute things from this dependency as part of the compilation.