refactor(toolchain): add deps_srcs to sdl rule (#4241)
* refactor(toolchain): change prod_tsc_build to not write hash to output

* fix(toolchain): use exec_dep when declaring generator

From https://buck2.build/docs/api/build/attrs/#attrsexec_dep:

Takes a target from the user, as a string, and supplies a dependency to the rule. The dependency will transition to the execution platform. Use exec_dep if you plan to execute things from this dependency as part of the compilation.

* chore(toolchain): add src optional dep to sdl task

* refactor(toolchain): change sdl srcs to 'deps_srcs' list type

* refactor(toolchain): add 'deps_srcs' arg to rust sdl rules

* fix(core): undo 'galoy_rust_bin' srcs change

Has to be passed as a glob directly because of hardcoded internal
details.

* chore(core): remove deps_srcs from rust sdl rules

* chore(core): revert filegroup addition to rust buck files

---------

Co-authored-by: bodymindarts <justin@galoy.io>